Fix result of parser resolve was cached
In getParser we return a function if the input is an array.
When called, this function resolves the value in the first
entry of the array and searches for a parser.
This resolving should be done each time this function is called,
but as we had written the result of the resolving to the variable
we resolved, there was nothing to resolve the next time the function
was called.

To fix this, we simply store the result of the resolving in a separate,
local variable